Split hooks into many files
