+++ src/components/Property/Types/VectorProperty/ViewOptions/DefaultView.tsx
@@ -1,6 +1,6 @@
 import { Flex } from '@mantine/core';
+import { usePropListeningState } from 'src/hooks/util';
 
-import { usePropListeningState } from '@/api/hooks';
 import { NumericInput } from '@/components/Input/NumericInput/NumericInput';
 import { AdditionalDataVectorMatrix } from '@/components/Property/types';
